Add: status, add: socket_service, update; main, update: pubspec
Add status into the pages folder, this page will contain the comunication between the server and the application.
Add socket service into the services folder, this page will contain the model to use provider and the connection of the client to the server.
Update main enclosing myApp into a multiprovider to create the instance of the provider in the higgest site of the app.
Update pubspec, adding provider and socket_io_client fron the pub.dev.