Update: socket_service, update: status
Update socket service adding a getter to the serverStatus to know the value when provider needs it, also in the socket.onConnect changing the value of the serverStatus to online and in onDisconnect changing serverStatus value to offline.
Update status adding a column to show the serverStatus in the middle of the screen with a text as its children that show the value of the serverStatus using provider.