Update: socket service
Update socket service adding the socket.on that prints a payload that is emit from the server, as the payload sends a known objetc that contains 'nombre' and 'mensaje' also making a conditional if mensaje2 is emit show 'No hay' but if its emit show the payload of mensaje2.
This code is to prepare the flutter application to listen the emits from the server, but its a test and will change in the future.