Update: main, update: home_page, update: status.
Update main changing the initial route from status to home.
Update home page adding the instance of SocketService through provider to put in the right corner an icon to show a check if the app is conected to the server or an offline wifi if is disconect.
Update status sending another test message to the server.